Revert "Move user schemas and modules into central datastore module"
This reverts commit 689786fa69ca4b6c01d413dee603516b5909ab74.
