+++ libsplinter/src/biome/user/store/mod.rs
@@ -18,8 +18,6 @@
 pub(in crate::biome) mod diesel;
 mod error;
 
-pub use crate::biome::datastore::models::UserModel;
-#[cfg(feature = "diesel")]
 pub use error::UserStoreError;
 
 /// Represents a user of a splinter application
@@ -44,13 +42,6 @@ impl SplinterUser {
     /// Returns the user's id.
     pub fn id(&self) -> String {
         self.id.to_string()
-    }
-}
-
-#[cfg(feature = "diesel")]
-impl From<UserModel> for SplinterUser {
-    fn from(user: UserModel) -> Self {
-        SplinterUser { id: user.id }
     }
 }
